Revert "Rétabli les link et utilise les routes parallèles pour /documents (#1…"
This reverts commit 12aedc4b29f0c4437d4a18f118e64083e0d98477.
