--- app/(header-default)/documents/[slug]/@carteProfessionnelleTP/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loader() {
-  return null;
-}
