--- app/(header-default)/documents/[slug]/@conformite/common.tsx
@@ -1,8 +1,0 @@
-import { EAdministration } from '#models/administrations/EAdministration';
-
-export const sectionInfo = {
-  title: 'Conformité',
-  id: 'conformite',
-  isProtected: true,
-  sources: [EAdministration.DGFIP, EAdministration.URSSAF, EAdministration.MSA],
-};
