--- components/title-section/tabs/TabLink.tsx
@@ -1,35 +1,0 @@
-'use client';
-import Link from 'next/link';
-import { useEffect } from 'react';
-import styles from './styles.module.css';
-type IProps = {
-  label: string;
-  href: string;
-  noFollow?: boolean;
-  width?: string;
-  active: boolean;
-};
-export default function TabLink({
-  active,
-  href,
-  label,
-  noFollow,
-  width,
-}: IProps) {
-  useEffect(() => {
-    active === false && window.dispatchEvent(new Event('cancelloadbar'));
-  }, [active]);
-  return (
-    <Link
-      className={`${active ? styles.activeLink : ''} no-style-link`}
-      href={href}
-      rel={noFollow ? 'nofollow' : ''}
-      style={{ width }}
-      scroll={false}
-      onClick={() => window.dispatchEvent(new Event('runloadbar'))}
-      prefetch={false}
-    >
-      {active ? label : <h2>{label}</h2>}
-    </Link>
-  );
-}
